physics system fix, physics now compiles
